Move and rename ReleasesMultigroup to nzedb\processing\ProcessReleasesMultiGroup
